requirements.txt
flask
tensorflow
Pillow
numpy